refactor: organize types
