+++ apps/calendar/src/components/dayGridMonth/gridCell.tsx
@@ -12,7 +12,6 @@ import {
 } from '@src/constants/style';
 import { useDispatch } from '@src/contexts/calendarStore';
 import { useTheme } from '@src/contexts/theme';
-import { Size } from '@src/controller/panel';
 import { cls, toPercent } from '@src/helpers/css';
 import { getExceedCount } from '@src/helpers/grid';
 import EventUIModel from '@src/model/eventUIModel';
@@ -22,6 +21,7 @@ import { Day } from '@src/time/datetime';
 import { getPosition, getRelativePosition, getSize } from '@src/utils/dom';
 import { ratio } from '@src/utils/math';
 
+import { RectSize } from '@t/layout';
 import { PopupRect } from '@t/store';
 
 interface Props {
@@ -45,7 +45,7 @@ function getSeeMorePopupSize(
   offsetWidth: number,
   eventLength: number,
   options: SeeMoreOptions
-): Size {
+): RectSize {
   const minHeight = getSize(grid).height + MONTH_MORE_VIEW_PADDING * 2;
   let width = offsetWidth + MONTH_MORE_VIEW_PADDING * 2;
 
@@ -81,8 +81,8 @@ function getSeeMorePopupSize(
 
 function getSeeMorePopupPosition(
   position: [positionX: number, positionY: number],
-  popupSize: Size,
-  appContainerSize: Size
+  popupSize: RectSize,
+  appContainerSize: RectSize
 ) {
   const { width: containerWidth, height: containerHeight } = appContainerSize;
   const [leftPos, topPos] = position;
