+++ apps/calendar/src/components/panelResizer.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent, h } from 'preact';
 
 import { Draggable, DragPositionInfo } from '@src/components/draggable';
-import { Direction } from '@src/controller/layout';
+import { Direction } from '@src/constants/layout';
 import { cls } from '@src/helpers/css';
 
 const BORDER = '1px solid #e5e5e5';
