+++ apps/calendar/src/constants/layout.ts
@@ -1,0 +1,34 @@
+import { PanelInfo } from '@t/layout';
+
+export enum Direction {
+  COLUMN,
+  ROW,
+}
+
+export enum ResizeMode {
+  RELATIVE,
+  ABSOLUTE,
+}
+
+export const panelInfoKeys: Array<keyof PanelInfo> = [
+  'name',
+  'direction',
+  'overflowX',
+  'overflowY',
+  'height',
+  'width',
+  'minHeight',
+  'minWidth',
+  'maxHeight',
+  'maxWidth',
+  'maxExpandableHeight',
+  'maxExpandableWidth',
+  'expandable',
+  'show',
+  'resizable',
+  'autoSize',
+  'resizerHeight',
+  'resizerWidth',
+];
+
+export const DEFAULT_PANEL_HEIGHT = 72;
