+++ apps/calendar/src/controller/panel.ts
@@ -1,65 +1,8 @@
-import { Direction } from '@src/controller/layout';
 import { getElementRect as getRect } from '@src/utils/dom';
 import { isUndefined } from '@src/utils/type';
 
-export type Styles = Record<string, string | number>;
-export type PanelElementRectMap = Record<string, PanelRect>;
-export interface Size {
-  width: number;
-  height: number;
-}
-export interface PanelRect extends Size {
-  x: number;
-  y: number;
-  resizerWidth: number;
-  resizerHeight: number;
-}
-export interface PanelInfo {
-  name: string;
-  direction?: Direction;
-  overflowY?: boolean;
-  overflowX?: boolean;
-  show?: boolean;
-
-  autoSize?: number;
-  height?: number;
-  width?: number;
-  minHeight?: number;
-  minWidth?: number;
-  maxHeight?: number;
-  maxWidth?: number;
-
-  expandable?: boolean;
-  maxExpandableHeight?: number;
-  maxExpandableWidth?: number;
-
-  resizable?: boolean;
-  resizerHeight?: number;
-  resizerWidth?: number;
-}
-
-export const panelInfoKeys: Array<keyof PanelInfo> = [
-  'name',
-  'direction',
-  'overflowX',
-  'overflowY',
-  'height',
-  'width',
-  'minHeight',
-  'minWidth',
-  'maxHeight',
-  'maxWidth',
-  'maxExpandableHeight',
-  'maxExpandableWidth',
-  'expandable',
-  'show',
-  'resizable',
-  'autoSize',
-  'resizerHeight',
-  'resizerWidth',
-];
-
-export const DEFAULT_PANEL_HEIGHT = 72;
+import { StyleProp } from '@t/components/common';
+import { PanelInfo, PanelRect } from '@t/layout';
 
 const styleKeys: Array<keyof PanelInfo> = ['minHeight', 'maxHeight', 'minWidth', 'maxWidth'];
 
@@ -80,7 +23,7 @@ function getPanelSide(side: number, maxExpandableSide?: number) {
 }
 
 export function getPanelStylesFromInfo(panel: PanelInfo) {
-  const styles: Styles = {};
+  const styles: StyleProp = {};
   const { height, width, overflowX, overflowY, maxExpandableWidth, maxExpandableHeight } = panel;
 
   if (width) {
