--- apps/calendar/src/event/externalEventType.ts
@@ -1,1 +1,0 @@
-export type ExternalEventName = 'beforeCreateEvent' | 'clickEvent';
