--- apps/calendar/src/event/internalEventType.ts
@@ -1,1 +1,0 @@
-export type InternalEventName = 'click' | 'dblclick' | 'render';
