+++ apps/calendar/stories/columns.stories.tsx
@@ -2,7 +2,6 @@ import { h, RenderableProps } from 'preact';
 
 import range from 'tui-code-snippet/array/range';
 
-import { GridSelectionInfo } from '@src/components/timeGrid';
 import { Column } from '@src/components/timeGrid/column';
 import { cls } from '@src/helpers/css';
 import TZDate from '@src/time/date';
@@ -15,6 +14,7 @@ import {
   toStartOfYear,
 } from '@src/time/datetime';
 
+import { TimeGridSelectionInfo } from '@t/components/timeGrid/gridSelection';
 import { EventModelData } from '@t/events';
 
 import { createEventModels } from '@stories/helper/event';
@@ -232,7 +232,7 @@ columnHasBackgroundColor.story = {
 
 export const gridSelectionWithTop = () => {
   const start = toStartOfDay(new TZDate());
-  const gridSelection: GridSelectionInfo = {
+  const gridSelection: TimeGridSelectionInfo = {
     start: addHours(start, 2),
     end: addHours(start, 5),
     unit: 'minute',
@@ -250,7 +250,7 @@ gridSelectionWithTop.story = {
 
 export const gridSelectionWithBottom = () => {
   const start = toStartOfDay(new TZDate());
-  const gridSelection: GridSelectionInfo = {
+  const gridSelection: TimeGridSelectionInfo = {
     start: addHours(start, 3),
     end: addHours(start, 6),
     unit: 'minute',
