+++ apps/calendar/stories/layout.stories.tsx
@@ -2,7 +2,7 @@ import { h } from 'preact';
 
 import { Layout } from '@src/components/layout';
 import Panel from '@src/components/panel';
-import { Direction, ResizeMode } from '@src/controller/layout';
+import { Direction, ResizeMode } from '@src/constants/layout';
 
 import { ProviderWrapper } from '@stories/util/providerWrapper';
