+++ apps/calendar/types/events.d.ts
@@ -1,7 +1,7 @@
-import EventModel from '@src/model/eventModel';
-import EventUIModel from '@src/model/eventUIModel';
-import TZDate from '@src/time/date';
-import Collection from '@src/utils/collection';
+import type EventModel from '@src/model/eventModel';
+import type EventUIModel from '@src/model/eventUIModel';
+import type TZDate from '@src/time/date';
+import type Collection from '@src/utils/collection';
 
 import { CalendarInfo } from '@t/option';
