+++ apps/calendar/types/template.d.ts
@@ -1,7 +1,7 @@
 import { VNode } from 'preact';
 
-import EventModel from '@src/model/eventModel';
-import TZDate from '@src/time/date';
+import type EventModel from '@src/model/eventModel';
+import type TZDate from '@src/time/date';
 
 import { EventCategory, TimeUnit } from '@t/events';
 import { TimezoneConfig } from '@t/option';
